refactor(app): split AppState responsibilities into submodules
