+++ view/app/src/app_state/settings_accessors/mod.rs
@@ -1,0 +1,5 @@
+//! AppState の設定アクセサ群を種類別に分割したモジュール。
+
+mod camera_settings;
+mod octree_settings;
+mod snapshot_settings;
